requirements.txt
# ── Core ─────────────────────────────────────────────────────────────────────
python-dotenv>=1.0
requests>=2.32
beautifulsoup4>=4.12
pytz>=2024.1
apscheduler>=3.10

# ── Telegram Bot ──────────────────────────────────────────────────────────────
python-telegram-bot>=21.0

# ── Market Data ───────────────────────────────────────────────────────────────
yfinance>=0.2.40
feedparser>=6.0

# ── AI / LLM ──────────────────────────────────────────────────────────────────
anthropic>=0.40
openai>=1.50          # deepseek 호환
groq>=0.11

# ── Google / YouTube ─────────────────────────────────────────────────────────
google-api-python-client>=2.130
google-auth>=2.30
google-auth-httplib2>=0.2
google-auth-oauthlib>=1.2

# ── PDF 처리 ──────────────────────────────────────────────────────────────────
pdfplumber>=0.11

# ── 영상 생성 ─────────────────────────────────────────────────────────────────
moviepy>=2.0
Pillow>=10.0

# ── AI 영상 생성 (Sora급 오픈소스) ───────────────────────────────────────────
fal-client>=0.10          # fal.ai SDK — Flux 이미지 / Wan2.1 text-to-video
edge-tts>=6.1             # Microsoft Edge TTS (한국어/영어 나레이션, 무료)

# ── Database (선택) ───────────────────────────────────────────────────────────
supabase>=2.4
